requirements.txt
streamlit
keras.models
datetime
yfinance
pandas_datareader
matplotlib.pyplot
numpy
pandas
streamlit_option_menu
sklearn
prophet
plotly